requirements.txt
python-telegram-bot==13.15

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
requests==2.31.0

sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

gspread==6.0.0
google-auth==2.27.0

redis==5.0.1

pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1